+++ modules/core/index.ts
@@ -1,0 +1,2 @@
+export { NodeRunnerResolverBase } from './resolver/node-runner.resolver';
+export { WorkerRunnerResolverBase } from './resolver/worker-runner.resolver';
