+++ modules/promise/dev/runner.resolver.ts
@@ -1,6 +1,6 @@
 import { RunnerConstructor } from '@core/types/constructor';
 import { WorkerBridge } from '@core/worker-bridge';
-import { NodeRunnerResolver } from '../node-runner.resolver';
+import { NodeRunnerResolver } from '../resolvers/node-runner.resolver';
 import { DevWorkerRunnerResolver } from './worker-runner.resolver';
 
 export class DevRunnerResolver<R extends RunnerConstructor> extends NodeRunnerResolver<R> {
