+++ modules/promise/dev/worker-runner.resolver.ts
@@ -1,7 +1,7 @@
 import { IWorkerAction } from '@core/actions/worker.actions';
 import { WorkerRunnerResolverBase } from '@core/resolver/worker-runner.resolver';
 import { Constructor } from '@core/types/constructor';
-import { DevWorkerBridge } from '@modules/promise/dev/dev-worker-bridge';
+import { DevWorkerBridge } from '@modules/promise/dev/worker-bridge';
 
 export class DevWorkerRunnerResolver<R extends Constructor<{[key: string]: any}>> extends WorkerRunnerResolverBase<R> {
     public workerBridge = new DevWorkerBridge(this);
