+++ modules/promise/index.ts
@@ -1,0 +1,1 @@
+export { RunnerResolver } from './resolvers/runner.resolver';
