+++ modules/promise/resolvers/node-runner.resolver.ts
