+++ modules/rx/dev/runner.resolver.ts
@@ -1,6 +1,6 @@
 import { RunnerConstructor } from '@core/types/constructor';
 import { RxNodeRunnerResolver } from '../resolvers/node-runner.resolver';
-import { RxDevWorkerBridge } from './dev-worker-bridge';
+import { RxDevWorkerBridge } from './worker-bridge';
 import { RxDevWorkerRunnerResolver } from './worker-runner.resolver';
 
 export class RxDevRunnerResolver<R extends RunnerConstructor> extends RxNodeRunnerResolver<R> {
