+++ modules/rx/index.ts
@@ -1,0 +1,1 @@
+export { RxRunnerResolver } from './resolvers/runner.resolver';
