+++ test/common/promise.ts
@@ -1,5 +1,5 @@
 import { DevRunnerResolver } from '@modules/promise/dev/runner.resolver';
-import { RunnerResolver } from '@modules/promise/runner.resolver';
+import { RunnerResolver } from '@modules/promise/resolvers/runner.resolver';
 import { runners } from './runner-list';
 
 export const runnerResolver = new RunnerResolver({runners, workerPath: 'base/test/worker.js'});
