+++ test/common/rx.ts
@@ -1,5 +1,5 @@
-import { RxDevRunnerResolver } from '@modules/rxjs/dev/runner.resolver';
-import { RxRunnerResolver } from '@modules/rxjs/resolvers/runner.resolver';
+import { RxDevRunnerResolver } from '@modules/rx/dev/runner.resolver';
+import { RxRunnerResolver } from '@modules/rx/resolvers/runner.resolver';
 import { runners } from './runner-list';
 
 export const rxResolver = new RxRunnerResolver({runners, workerPath: 'base/test/rx-worker.js'});
