+++ tsconfig.json
@@ -8,8 +8,9 @@
         "target": "ES2020",
         "outDir": "./dist",
         "strictNullChecks": true,
-        "strict": true,        
+        "strict": true,
         "paths": {
+            "@core": ["./modules/core"],
             "@core/*": ["./modules/core/*"],
             "@modules/*": ["./modules/*"],
         }
